moved pix functions into its own package
