+++ mysite/ratata/pix/__init__.py
@@ -1,0 +1,1 @@
+from .pix import *
