+++ mysite/ratata/templates/ratata/components/pix.html
@@ -1,7 +1,10 @@
 <div hx-get="/ratata/accounts/{{account.id}}/payments/{{user.id}}" hx-trigger="newTransaction from:body">
     {% if pix %}
-		<img src="data:image/png;base64,{{pix.image}}">
-		<button class="btn btn-primary pix-qr" onclick="copy()">Copy</button>
+		Pague com pix:
+		<div class="flex-col">
+			<img class="pix-img" src="data:image/png;base64,{{pix.image}}">
+			<button class="btn btn-primary pix-qr" onclick="copy()">Copiar código</button>
+		</div>
     {% endif %}
 </div>
 <script>
